For_GitHub_Upload/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pdfplumber==0.10.3
python-multipart==0.0.6
gunicorn==21.2.0
werkzeug==3.0.0
requests==2.31.0
flask-limiter==3.5.0
